+++ resolwe/flow/execution_engines/bash/__init__.py
@@ -65,7 +65,7 @@ class ExecutionEngine(BaseExecutionEngine):
             inputs["requirements"] = data.process.requirements
             # Inject default values and change resources according to
             # the current Django configuration.
-            inputs["requirements"]["resources"] = data.process.get_resource_limits()
+            inputs["requirements"]["resources"] = data.get_resource_limits()
 
             script_template = data.process.run.get("program", "")